# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UK Education & Career Guide API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type SubjectLevel = (typeof SubjectLevel)[keyof typeof SubjectLevel];

export const SubjectLevel = {
  GCSE: "GCSE",
  "A-Level": "A-Level",
  Both: "Both",
} as const;

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType];

export const ResourceType = {
  Website: "Website",
  Video: "Video",
  Past_Papers: "Past Papers",
  Revision_Guide: "Revision Guide",
  Practice_Questions: "Practice Questions",
} as const;

export interface Resource {
  name: string;
  url?: string;
  type: ResourceType;
}

export interface Subject {
  id: number;
  name: string;
  level: SubjectLevel;
  category: string;
  description: string;
  examBoards?: string[];
  studyTips?: string[];
  keyTopics?: string[];
  assessmentStructure?: string;
  usefulResources?: Resource[];
  relatedCareers?: number[];
}

export type CareerJobOutlook =
  (typeof CareerJobOutlook)[keyof typeof CareerJobOutlook];

export const CareerJobOutlook = {
  Excellent: "Excellent",
  Good: "Good",
  Moderate: "Moderate",
  Declining: "Declining",
} as const;

export interface Career {
  id: number;
  title: string;
  sector: string;
  description: string;
  averageSalaryMin: number;
  averageSalaryMax: number;
  requiredSubjects?: number[];
  preferredSubjects?: number[];
  entryRoutes?: string[];
  requiredQualifications?: string[];
  jobOutlook?: CareerJobOutlook;
  workplaces?: string[];
  dayInTheLife?: string;
  relatedCareers?: number[];
}

export type InstitutionType =
  (typeof InstitutionType)[keyof typeof InstitutionType];

export const InstitutionType = {
  University: "University",
  College: "College",
  Apprenticeship_Provider: "Apprenticeship Provider",
  Conservatoire: "Conservatoire",
  Specialist: "Specialist",
} as const;

export interface Institution {
  id: number;
  name: string;
  type: InstitutionType;
  region: string;
  city: string;
  description: string;
  websiteUrl?: string;
  russellGroup?: boolean;
  ranking?: number;
  entryRequirements?: string;
  ucasPoints?: number;
  studentSatisfaction?: number;
  notableSubjects?: string[];
  facilities?: string[];
  bursaries?: boolean;
  internationalStudents?: boolean;
  relatedCareers?: number[];
}

export type RouteAfterLevel =
  (typeof RouteAfterLevel)[keyof typeof RouteAfterLevel];

export const RouteAfterLevel = {
  GCSE: "GCSE",
  "A-Level": "A-Level",
} as const;

export interface Route {
  id: number;
  name: string;
  type: string;
  afterLevel: RouteAfterLevel;
  duration: string;
  description: string;
  pros?: string[];
  cons?: string[];
  entryRequirements?: string;
  potentialEarnings?: string;
  examples?: string[];
}

export type QuizInputGcseSubjectsItem = {
  subjectId: number;
  grade: string;
};

export type QuizInputALevelSubjectsItem = {
  subjectId: number;
  predictedGrade: string;
};

export type QuizInputPreferredRouteType =
  (typeof QuizInputPreferredRouteType)[keyof typeof QuizInputPreferredRouteType];

export const QuizInputPreferredRouteType = {
  University: "University",
  Apprenticeship: "Apprenticeship",
  College: "College",
  Any: "Any",
} as const;

export interface QuizInput {
  gcseSubjects?: QuizInputGcseSubjectsItem[];
  aLevelSubjects?: QuizInputALevelSubjectsItem[];
  interests: string[];
  preferredRegion?: string;
  preferredRouteType?: QuizInputPreferredRouteType;
}

export interface QuizResult {
  recommendedCareers: Career[];
  recommendedInstitutions: Institution[];
  recommendedRoutes: Route[];
  matchScore?: number;
  personalMessage: string;
}

export type GetSubjectsParams = {
  level?: GetSubjectsLevel;
  category?: string;
};

export type GetSubjectsLevel =
  (typeof GetSubjectsLevel)[keyof typeof GetSubjectsLevel];

export const GetSubjectsLevel = {
  GCSE: "GCSE",
  "A-Level": "A-Level",
  Both: "Both",
} as const;

export type GetCareersParams = {
  sector?: string;
  subjectId?: number;
  minSalary?: number;
};

export type GetInstitutionsParams = {
  type?: GetInstitutionsType;
  region?: string;
  russellGroup?: boolean;
  minEntryGrade?: string;
  careerId?: number;
};

export type GetInstitutionsType =
  (typeof GetInstitutionsType)[keyof typeof GetInstitutionsType];

export const GetInstitutionsType = {
  University: "University",
  College: "College",
  Apprenticeship_Provider: "Apprenticeship Provider",
  Conservatoire: "Conservatoire",
  Specialist: "Specialist",
} as const;

export type GetRoutesParams = {
  afterLevel?: GetRoutesAfterLevel;
  type?: string;
};

export type GetRoutesAfterLevel =
  (typeof GetRoutesAfterLevel)[keyof typeof GetRoutesAfterLevel];

export const GetRoutesAfterLevel = {
  GCSE: "GCSE",
  "A-Level": "A-Level",
} as const;