# lib/api-zod/src/generated/types/institution.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UK Education & Career Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { InstitutionType } from "./institutionType";

export interface Institution {
  id: number;
  name: string;
  type: InstitutionType;
  region: string;
  city: string;
  description: string;
  websiteUrl?: string;
  russellGroup?: boolean;
  ranking?: number;
  entryRequirements?: string;
  ucasPoints?: number;
  studentSatisfaction?: number;
  notableSubjects?: string[];
  facilities?: string[];
  bursaries?: boolean;
  internationalStudents?: boolean;
  relatedCareers?: number[];
}