# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UK Education & Career Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizInputALevelSubjectsItem } from "./quizInputALevelSubjectsItem";
import type { QuizInputGcseSubjectsItem } from "./quizInputGcseSubjectsItem";
import type { QuizInputPreferredRouteType } from "./quizInputPreferredRouteType";

export interface QuizInput {
  gcseSubjects?: QuizInputGcseSubjectsItem[];
  aLevelSubjects?: QuizInputALevelSubjectsItem[];
  interests: string[];
  preferredRegion?: string;
  preferredRouteType?: QuizInputPreferredRouteType;
}