# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UK Education & Career Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from "./resourceType";

export interface Resource {
  name: string;
  url?: string;
  type: ResourceType;
}