# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UK Education & Career Guide API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteAfterLevel } from "./routeAfterLevel";

export interface Route {
  id: number;
  name: string;
  type: string;
  afterLevel: RouteAfterLevel;
  duration: string;
  description: string;
  pros?: string[];
  cons?: string[];
  entryRequirements?: string;
  potentialEarnings?: string;
  examples?: string[];
}